pyproject.toml
[tool.poetry]
name = "backstriper"
version = "0.1.0"
description = ""
authors = ["Helmut Irle <helmut@devigninnovation.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
typer = {extras = ["all"], version = "^0.9.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
